improv(api): remove the accept-first feature
It assumes contextual information (one cannot look at the target
blindly), which is against the very aim of Leap. The case when the
target is on a different line (= cannot be targeted with f/t), but
it's still 100% obvious that there's no other target in between, is
probably too rare to make this feature worthwhile.