improv: highlight matches only in traversal mode
Match highlights are actually redundant - so why not get rid of them
when they are not necessary, to further simplify the UI?

Caveats:

An unlabeled match might cover a label - in that case, the highlight
will be turned on, to make the user aware of the conflict.

Since unhighlighted pairs indicate directly reachable matches now, even
inactive labels should be displayed, at least as empty blocks.