refactor(leap): recurse with `traverse` instead of `leap` itself
So obvious in hindsight, simplifies the function signature and lots of
other things. A nice side-effect is that the `jump-to!` closure is not
destroyed when entering traversal mode (the `first-jump?` variable is
kept maintained).

`disable-labels` is renamed to `inactivate-labels`, and sets the state
to :inactive instead of nil. (The very first target might have no label
state, so nil is not a good short-circuit condition for
`get-target-with-active-primary-label`.)