improv: resolving phase one conflicts
In case of labeled-unlabeled conflicts, we choose the simplest, most
consistent solution (at least UI-wise): label the unlabeled match too
(that unfortunately implies looping through the whole sublist again, and
re-labeling all the items).

Note: As even a unique target can get a label now in
`resolve-conflicts`, in the main algorithm we should check for a label
before hastily doing the action and exiting.