improv(defaults): new default mappings
After due consideration, added a new method `create_default_mappings`,
that only maps `s` (as exclusive motion) and `S`, besides `gs`.

"One obvious way to do things", for UIs at least, is usually preferable
if there are no strong arguments for doing otherwise. This new
arrangement is easier to document and remember, and can more easily
become muscle memory, even if in some cases it might prove to be a bit
less comfortable than the old way.

- `vs` (forced inclusive) provides enough flexibility, given that Leap
  can search for newline characters, one can still select any visible
  region for an operation.

- `X` was never too intuitive in the first place, as you need to start
  with a character that is not even on the border of the intended
  selection (it's a bit like having to type the pattern backwards for
  `x`). And we still have the old `x`, in the form of `s` now.

- The behaviors of `vs` (including the first character) and the old `s`
  (including both characters) are roughly equally intuitive, it depends
  on the specific situation. The latter was handy of course, but we can
  live without it.

Needless to say, it is a huge win that we are not taking over another
pair of native (even though redundant) Vim mappings.