Improve an example of linewise motions
Extend the example of linewise motions to include the ability to jump to
the start of the line or to try to keep the column after the jump.

Also add and option and mappings for the directional jumps (upwards and
downwards).

Create keymaps for "|", "<leader><leader>J", "<leader><leader>K",
"<leader><leader>j", "<leader><leader>k" in normal, visual, and
operator-pending modes.