improv!: simplify UI (all labels are created equal)
BREAKING CHANGE

Follow-up to 2ec33f2a38974fe6997c42c5ce61ce84eab84cc8.

Labels in subsequent groups are always hidden from now on. Consequently,
we're introducing a new highlight group, `LeapLabel`, to be used for any
label, replacing `LeapLabelPrimary` and `LeapLabelSecondary`.

`LeapLabelPrimary` is kept as an alias for `LeapLabel`, but deprecated,
and should be renamed if possible.

Rationale:

Having to focus on two separate attributes of the label makes the UI a
tad bit too complicated, even in multi-phase mode. Furthermore, relying
solely on the highlighting style to signal inactivity is dangerous; the
two highlight groups should be clearly distinguishable, a requirement
easily messed up by color schemes, which we have no (direct) control
over.

We should optimize for the common case, and make it as straightforward
as possible. Using a two-character search pattern and an extended label
set makes group switching an exceptional case in this plugin, and it
should simply be handled as such, without making too big a fuss about
it.

Alternatives considered:

1. Using yet another, configurable list for "group labels".

Too complicated, for little gain. We would need a new `opts` field, and
ideally a separate highlight group (to signal that this is not the final
label yet).

2. Using uppercase letters (A-Z) for group labels.

Hurts the common case a lot, as we lose dozens of letters that could be
used as normal labels instead.

3. Using numbers (1-9) for group labels.

Seems a pretty neat solution. The numbers can be used directly as
labels, but they also show how many times to press the switch key, if
one prefers to use that instead. This can be added anytime in the future
though, without API change.