improv(defaults)!: `offset` chages in -till/-to keys
BREAKING CHANGE

Backward search with offset=2 is a bit like having to type backwards
from the second match character when doing forward search - awkward. An
"exclusive" motion pair can be useful, but it is more intuitive to
consistently treat the first match character as the reference point, and
the second one as merely a filter for search, in all motions. (For the
same reason, the offset argument from `<Plug>(leap-forward-to)` has been
removed.)